垃圾收集器之引用计数算法示例

+++ src/main/java/com/zxb/understanding/the/jvm/oom/RuntimeConstantPoolOom.java
@@ -7,7 +7,8 @@ import java.util.List;
  * {@link OutOfMemoryError} 方法区和运行时常量池溢出
  * VM Args: -XX:PermSize=5m -XX:MaxPermSize=5m
  * JDK1.6 会出现 java.lang.OutOfMemoryError: PermGe space
- * Jdk1.7 会一直运行下去，需要加 -Xmx5m，因为Jdk7以上已经将常量池移至Java堆中
+ * Jdk1.7 会一直运行下去，需要加 -Xmx5m，才会出现 java.lang.OutOfMemoryError: Java heap space
+ * 可以看出因为Jdk7以上已经将常量池移至Java堆中
  *
  * @author Mr.zxb
  * @date 2020-01-16 15:49
@@ -19,8 +20,8 @@ public class RuntimeConstantPoolOom {
         // 使用list 保持着常量池引用，避免 Full GC 回收常量池行为
         List<String> list = new ArrayList<>();
 
-        // 10M 的 PermSize 在 Integer 范围内足够产生 OOM了
-        int i = 0;
+        // 在short范围内足以让 10M 的 PermSize产生 OOM了
+        short i = 0;
         while (true) {
             list.add(String.valueOf(i++).intern());
         }
